CBD-4364: Handle missing project directories
If a directory is missing but was expected to be missing (ie., it was
skipped by "repo init -g" filters, or isn't in the manifest at all),
that's OK - we assume that whoever synced the manifest is asserting
the set of projects they chose is self-contained, so changes to other
projects aren't required.

If a directory is missing and this was NOT expected, raise an error.

Change-Id: I38085f0936c9f0bed5e0d552f40c21c761fdc6b8
Reviewed-on: http://review.couchbase.org/c/patch_via_gerrit/+/162252
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>